add orderManager and orderView

+++ Shkaf/Shkaf/Models/Product.swift
@@ -13,14 +13,14 @@ struct Product: Identifiable {
     var name: String
     var description: String
     var image: String
-    var price: Int
+    var price: Double
 }
 
-var productList = [Product(categoty: "Одежда", name: "Orange sweater", description: "Свитер", image: "sweater1", price: 121),
-                   Product(categoty: "Одежда", name: "Red sweater", description: "Свитер", image: "sweater2", price: 99),
-                   Product(categoty: "Одежда", name: "Blue sweater", description: "Свитер", image: "sweater3", price: 45),
-                   Product(categoty: "Одежда", name: "Green sweater", description: "Свитер", image: "sweater4", price: 89),
-                   Product(categoty: "Одежда", name: "Brown sweater", description: "Свитер", image: "sweater5", price: 201),
-                   Product(categoty: "Одежда", name: "Yellow sweater", description: "Свитер", image: "sweater6", price: 23),
-                   Product(categoty: "Одежда", name: "White sweater", description: "Свитер", image: "sweater7", price: 48),
-                   Product(categoty: "Одежда", name: "Cream sweater", description: "Свитер", image: "sweater8", price: 98)]
+var productList = [Product(categoty: "Одежда", name: "Orange sweater", description: "Свитер", image: "sweater1", price: 121.0),
+                   Product(categoty: "Одежда", name: "Red sweater", description: "Свитер", image: "sweater2", price: 99.0),
+                   Product(categoty: "Одежда", name: "Blue sweater", description: "Свитер", image: "sweater3", price: 45.0),
+                   Product(categoty: "Одежда", name: "Green sweater", description: "Свитер", image: "sweater4", price: 89.0),
+                   Product(categoty: "Одежда", name: "Brown sweater", description: "Свитер", image: "sweater5", price: 201.0),
+                   Product(categoty: "Одежда", name: "Yellow sweater", description: "Свитер", image: "sweater6", price: 23.0),
+                   Product(categoty: "Одежда", name: "White sweater", description: "Свитер", image: "sweater7", price: 48.0),
+                   Product(categoty: "Одежда", name: "Cream sweater", description: "Свитер", image: "sweater8", price: 98.0)]
